% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{fill_defaults}
\alias{fill_defaults}
\title{Fill in sensible default priors for a given p_family,
then override with any user-supplied priors.}
\usage{
fill_defaults(priors = list(), p_family, model_type = "glm")
}
\arguments{
\item{priors}{A named list of user-specified priors, e.g.
list(beta1 = "normal(0,5)", theta = "beta(2,2)")}

\item{p_family}{One of "linear", "poisson", "logistic", or "gamma"}

\item{model_type}{One of 'glm' (generalized linear model) or 'survival'
(survival model). By default, 'glm'.}
}
\value{
A named list of priors in the same format as \code{priors},
containing one entry for every required parameter.
}
\description{
Fill in sensible default priors for a given p_family,
then override with any user-supplied priors.
}
\keyword{internal}
