% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_helpers.R
\name{generate_stan}
\alias{generate_stan}
\title{Generate stan code for mixture models}
\usage{
generate_stan(components, formula, data)
}
\arguments{
\item{components}{Vector of component types}

\item{formula}{Model formula}

\item{data}{Input data}
}
\value{
Path to generated Stan file
}
\description{
Generate stan code for mixture models
}
\keyword{internal}
