% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_helpers.R
\name{generate_synthetic_mixture_data}
\alias{generate_synthetic_mixture_data}
\title{Creates synthetic data for testing mixture models.}
\usage{
generate_synthetic_mixture_data(family, seed)
}
\arguments{
\item{family}{Distribution family ("linear", "logistic", "poisson", "gamma")}

\item{seed}{Random seed}
}
\value{
A data frame with synthetic data
}
\description{
Creates synthetic data for testing mixture models.
}
\keyword{internal}
