% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_helpers.R
\name{generate_synthetic_survival_data}
\alias{generate_synthetic_survival_data}
\title{Creates synthetic data for testing survival models.}
\usage{
generate_synthetic_survival_data(family, seed)
}
\arguments{
\item{family}{Distribution family ("weibull", "gamma")}

\item{seed}{Random seed}
}
\value{
A data frame with synthetic data
}
\description{
Creates synthetic data for testing survival models.
}
\keyword{internal}
