% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_helpers.R
\name{get_mixture_results}
\alias{get_mixture_results}
\title{Process model results}
\usage{
get_mixture_results(family, fit, return_type)
}
\arguments{
\item{family}{Distribution family}

\item{fit}{Stan fit object}

\item{return_type}{0 for matrix output, 1 for posterior samples}
}
\value{
Processed results
}
\description{
Process model results
}
\keyword{internal}
