% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_helpers.R, R/survival_helpers.R
\name{get_results}
\alias{get_results}
\title{Process model results}
\usage{
get_results(family, fit, return_type = 0, status_vector = NULL)

get_results(family, fit, return_type = 0, status_vector = NULL)
}
\arguments{
\item{family}{Distribution family}

\item{fit}{Stan fit object}

\item{return_type}{0 for matrix output, 1 for posterior samples}

\item{status_vector}{vector for status column}
}
\value{
Processed results

Processed results
}
\description{
Process model results

Process model results
}
\keyword{internal}
