% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{get_stan_definitions}
\alias{get_stan_definitions}
\title{Takes a list of priors and uses process_variable() on them to generate strings
for concatenation into the stan code to be generated.}
\usage{
get_stan_definitions(priors)
}
\arguments{
\item{priors}{The named list of priors and variable definitions}
}
\value{
A list with keys varaible_defs and function_defs where the corresponding
values are the strings to be directly inserted into the stan code
}
\description{
Takes a list of priors and uses process_variable() on them to generate strings
for concatenation into the stan code to be generated.
}
