% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helpers.R
\name{get_stats}
\alias{get_stats}
\title{process_results helper that takes a numeric vector of draws for a parameter
component (a real parameter itself, or an element of a array parameter, etc.)
and returns summary statistics as a vector values}
\usage{
get_stats(v)
}
\arguments{
\item{v}{A numeric vector of draws. This is part of a parameter associated with a
single response variable (plus a single predictor in case of betas)}
}
\value{
A vector of statistics formatted as
c(\if{html}{\out{<mean>}}, \if{html}{\out{<variance>}}, <lower 95>, <upper 95>, <lower 50>, <upper 50>)
where lower 95 is the lower bound of the 95\% credible interval,
upper 50 is the upper bound of the 50\% credible interval, etc.
}
\description{
process_results helper that takes a numeric vector of draws for a parameter
component (a real parameter itself, or an element of a array parameter, etc.)
and returns summary statistics as a vector values
}
\keyword{internal}
