% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{is_func}
\alias{is_func}
\title{Checks if a prior list element was passed in as a stan function; used for
validation and prior list processing}
\usage{
is_func(stan_function)
}
\arguments{
\item{stan_function}{}
}
\value{
true if argument was constructed as stan_func("..."), false otherwise
}
\description{
Checks if a prior list element was passed in as a stan function; used for
validation and prior list processing
}
