% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{is_valid_stan_func}
\alias{is_valid_stan_func}
\title{Helper function that lightly validates a string as (intended to be) a stan
function. Specifically, returns true if str is of the form
...(...)...{...return...}... where '...' is any arbitrary string AND the
string does not contain R function assignment syntax '<-'}
\usage{
is_valid_stan_func(str)
}
\description{
Helper function that lightly validates a string as (intended to be) a stan
function. Specifically, returns true if str is of the form
...(...)...{...return...}... where '...' is any arbitrary string AND the
string does not contain R function assignment syntax '<-'
}
