% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_helpers.R
\name{LinksMixtureModeling}
\alias{LinksMixtureModeling}
\title{Fits Bayesian mixture models using Stan with various distribution families (package loader)}
\description{
Fits Bayesian mixture models using Stan with various distribution families (package loader)
}
\keyword{"_PACKAGE"}
\keyword{internal}
