% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helpers.R
\name{process_results}
\alias{process_results}
\title{Process model results from fit object and generate summaries}
\usage{
process_results(family, posterior, formulas)
}
\arguments{
\item{family}{Distribution family}

\item{posterior}{Posterior extracted from fit object returned by stampler}

\item{formulas}{List of formulas used on the stan model}
}
\value{
Processed results depending on return_type
}
\description{
Process model results from fit object and generate summaries
}
\keyword{internal}
