% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{process_variable}
\alias{process_variable}
\title{Processes a given key-value pair from the priors list into syntactically
valid stan to be inserted during stan generation. Helper function to
get_stan_definitions}
\usage{
process_variable(value, key)
}
\arguments{
\item{value}{Value ex. "normal(0,5)"}

\item{key}{key ex. beta1}
}
\value{
a list that organizes stan code by definition, declaration, and/or
stan function so it is apparent where the stan code should be inserted
}
\description{
Processes a given key-value pair from the priors list into syntactically
valid stan to be inserted during stan generation. Helper function to
get_stan_definitions
}
\keyword{internal}
