% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{stan_func}
\alias{stan_func}
\title{Wraps a string element in priors list to indicate that the string is intended
as a stan function.This string will be directly injected into a stan function
block with little validation and is intended for niche or advanced use cases.
Custom functions allow the definition of an arbitrary prior.}
\usage{
stan_func(stan_function_str)
}
\arguments{
\item{stan_function_str}{A string defining a function in stan syntax}
}
\value{
Inputted string wrapped to indicate the string as a stan function
injection
}
\description{
Wraps a string element in priors list to indicate that the string is intended
as a stan function.This string will be directly injected into a stan function
block with little validation and is intended for niche or advanced use cases.
Custom functions allow the definition of an arbitrary prior.
}
\examples{
lin_fit <- LinksMixtureModeling::fit_model(
  ... # other arguments here
  priors = list(
  my_function = stan_fun("
    real my_function(...) { 
      ... // function body
      return ... ; // return a prior distribution
    }
  "), # this function is injected into stan's function {} block
  theta1 = "my_function(...)" # reference function for other definitions
  )
  ... # other arguments here
)

}
