% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_helpers.R
\name{summary_template}
\alias{summary_template}
\title{process_results helper that creates data frame template for summary statistics}
\usage{
summary_template(row_names, row_count)
}
\arguments{
\item{num_of_preds}{Total number of rows for which summary being made}

\item{preds}{Vector of row names}
}
\value{
Empty data frame with named columns
}
\description{
process_results helper that creates data frame template for summary statistics
}
\keyword{internal}
