% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors_helpers.R
\name{validate_args}
\alias{validate_args}
\title{Initially validate user-supplied arguments; further validation occurs during
iteration over list elements in process_Variables. Primarily verifies
format for prior string follows \if{html}{\out{<dist>}}(\if{html}{\out{<comma separated args>}}) where
each arg is either a number or defined in the priors list.}
\usage{
validate_args(priors, p_family)
}
\arguments{
\item{priors}{A named list of Stan‐style prior strings, e.g.
list(mu1 = "normal(0,5))", sigma1 = "cauchy(0,2.5))}

\item{p_family}{One of "linear", "poisson", "logistic", or "gamma"}
}
\value{
Invisibly TRUE if all checks pass or only raise warning; errors otherwise.
}
\description{
Initially validate user-supplied arguments; further validation occurs during
iteration over list elements in process_Variables. Primarily verifies
format for prior string follows \if{html}{\out{<dist>}}(\if{html}{\out{<comma separated args>}}) where
each arg is either a number or defined in the priors list.
}
\keyword{internal}
