% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_survival_model.R
\name{fit_survival_model}
\alias{fit_survival_model}
\title{Fits a survival model using Stan with various distribution families.}
\usage{
fit_survival_model(
  formula,
  p_family,
  data,
  priors = NULL,
  result_type,
  iterations,
  burning_iterations,
  chains,
  seed,
  truncation = 0,
  status_column = NULL
)
}
\arguments{
\item{formula}{Model formula (e.g., y ~ x1 + x2)}

\item{p_family}{Distribution family ("weibull", "gamma")}

\item{data}{Input data frame or "random" for synthetic data}

\item{priors}{Named list (or NULL) of strings defining priors and optionally
hyperparameter definitions used in the prior definition. NULL priors or missing
list elements defining needed priors triggers weakly-informative defaults.}

\item{result_type}{0 for matrix output, 1 for posterior samples,}

\item{iterations}{Total number of MCMC iterations, default is}

\item{burning_iterations}{Number of burn-in iterations}

\item{chains}{Number of MCMC chains}

\item{seed}{Random seed (integer or "random")}

\item{truncation}{0 or 1 indicating whether to use truncated distributions}

\item{status_column}{indicates which column is the status one}
}
\value{
Results depending on result_type
}
\description{
Fits a survival model using Stan with various distribution families.
}
\examples{
\dontrun{
# Weibull mixture example
surv_fit <- LinksMixtureModeling::fit_survival_model(
formula = Surv(y, status) ~ X2, 
p_family = "weibull",
data = "random",
priors = list(beta1_sigma = cov_matrix, 
              beta1_loc = c(7,8), 
              beta1 = "multi_normal(beta1_loc, beta1_sigma)",
              a_func = stan_func("real mu_prior_lpdf(real x) { return normal_lpdf(x | 0, 5); }"),
              beta1 = "mu_prior()",
              beta2_loc = 0,
              beta2_scale = 6,
              beta2 = "normal(beta2_loc, beta2_scale)"),
result_type = 1,
iterations = 500,
burning_iterations = 250,
chains = 2,
seed = "random",
truncation = 0,
status_column = "status") }
}
